requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
reportlab==4.0.7
lxml==4.9.3
Pillow==10.1.0 